implemented rug crate for very large integers

+++ src/elliptic_curve/src/main.rs
@@ -1,4 +1,5 @@
 use elliptic_curve::EllipticPoint;
+use rug::Integer;
 
 fn main() {
     // Invalid
@@ -6,17 +7,32 @@ fn main() {
     // println!("Point 1 is valid: {}", point_1.is_valid());
 
 
-    let point_2 = EllipticPoint::new(Some(-1), Some(-1), 5, 7);
+    let point_2 = EllipticPoint::new(
+        Some(Integer::from(-1)),
+         Some(Integer::from(-1)), 
+         Integer::from(5), 
+        Integer::from(7)
+    );
     println!("Point 1 is valid: {}", point_2.is_valid());
 
 
-    let point_3 = EllipticPoint::new(Some(18), Some(77), 5, 7);
+    let point_3 = EllipticPoint::new(
+        Some(Integer::from(18)), 
+        Some(Integer::from(77)), 
+        Integer::from(5), 
+       Integer::from(7)
+    );
     println!("Point 1 is valid: {}", point_3.is_valid());
 
     // Invalid
     // let point_4 = EllipticPoint::new(Some(5), Some(7), 5, 7);
     // println!("Point 1 is valid: {}", point_4.is_valid());
 
-    let point_a = EllipticPoint::new(Some(-1), Some(-1), 5, 7);
+    let point_a = EllipticPoint::new(
+        Some(Integer::from(-1)),
+         Some(Integer::from(-1)), 
+         Integer::from(5), 
+        Integer::from(7)
+    );
     println!("Poin A + Point A: {:?}", point_a.clone() + point_a);
 }
